+++ include/render/video_renderer.h
@@ -46,6 +46,8 @@ public:
     virtual bool consumeSetABRepeatEndRequest() = 0;
     virtual bool consumeClearABRepeatRequest() = 0;
     virtual bool consumeScreenshotRequest() = 0;
+    virtual bool consumeStepFrameBackwardRequest() = 0;
+    virtual bool consumeStepFrameForwardRequest() = 0;
     virtual bool consumeNextChapterRequest() = 0;
     virtual bool consumePreviousChapterRequest() = 0;
     virtual bool consumeNextItemRequest() = 0;